moved experimental heuristics to its own file
